server/requirements.txt
flask>=2.0.2
bcrypt>=3.2.0
mysql-connector-python